+++ fm2prof/Fm2ProfRunner.py
@@ -15,7 +15,7 @@ from scipy.spatial import ConvexHull
 from fm2prof import Functions as FE
 from fm2prof import __version__
 from fm2prof.common import FM2ProfBase
-from fm2prof.CrossSection import CrossSection, CrossSectionHelpers
+from fm2prof.cross_section import CrossSection, CrossSectionHelpers
 from fm2prof.Export import Export1DModelData, OutputFiles
 from fm2prof.Import import FMDataImporter, FmModelData, ImportInputFiles
 from fm2prof.ini_file import IniFile
